stop automatically loading plugins in `getSupportInfo`
